shopblocksConverter.py created
updated cShop to now operate correctly with some minor tweaks per sheet. This will be altered down the line. Furthermore begun work on shopblocks. Set up a dispatch table and in an attempt to improve my programming ability I will be simplifying the original approach as much as possible for readibility and future extendability